ref-filter: make expand_data global
Put struct expand_data into global scope to reuse it
in cat-file.
